created base q-learning algorithm for the future development
